% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom.bigA.R
\name{dom.bigA}
\alias{dom.bigA}
\title{dom.bigA}
\usage{
dom.bigA(p,K,r,gens)
}
\arguments{
\item{p}{an object describing the initial frequency of A}

\item{K}{describes the carrying capacity}

\item{r}{a number giving the population growth rate}

\item{gens}{is the number of generations}
}
\description{
dom.bigA, a function describing selection on a population given a single climate change event that affects temperature by a 2 degree threshold. Exploring with the geontypes AA, Aa, and aa. A allele has dominance (wAa=wAA). This function estimates population size.
}
\examples{
K<-1000
p<-1/(2*K)
dom.bigA(p,K,2,100)

}
\references{
Bay, R. A., Rose, N. H., Logan, C. A., & Palumbi, S. R. (2017). Genomic models predict successful coral adaptation if future ocean warming rates are reduced. Science advances, 3(11), e1701413.

Gomulkiewicz, R., & Holt, R. D. (1995). When does evolution by natural selection prevent extinction?. Evolution, 49(1), 201-207.

Lindsey, H. A., Gallie, J., Taylor, S., & Kerr, B. (2013). Evolutionary rescue from extinction is contingent on a lower rate of environmental change. Nature, 494(7438), 463.
}
\author{
Jonathan Dickey
}
